pyproject.toml
[project]
name = "technician-scheduler"
version = "0.1.0"
description = "A comprehensive technician scheduling web application"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.0",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.2",
    "gunicorn>=23.0.0",
    "markupsafe>=3.0.2",
    "openpyxl>=3.1.5",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "pytz>=2024.2",
    "requests>=2.32.3",
    "sendgrid>=6.11.0",
    "sqlalchemy>=2.0.36",
    "trafilatura>=2.0.0",
    "werkzeug>=3.1.3",
    "wtforms>=3.2.1",
]
